Fix all warnings

+++ include/gestures.h
@@ -94,7 +94,7 @@ struct Gestures {
 	int scroll_speed_valid;
 
 	/* Duration left */
-	int coasting_duration_left;
+	float coasting_duration_left;
 };
